Work around for doctest issue with empty line
Sphinx does not work well when docstring output contains an empty line.
Since `@code_warntype` has a empty line in the printed output this was
causing Sphinx to have incorrect expected data. As it turns out the
docstring example is only interested in the return type of the method
which meant we could truncate part of the output using ... and avoid the
empty line issue.