Sort versions without build after versions with build
Currently we have v"0.5.0-dev" < v"0.5.0-dev+1". This breaks (at least) Compat for Julia installs where no build information is available.

I propose to change how versions are compared: If no build information is available, assume it is the very latest build, instead of assuming it is the oldest build.