store intermediate inference results in a separate table that's not saved
When inference is invoked on a call site we often only need the result type
and not the full code (in this case needtree==false is passed).
Store such results in a separate table, which is then not saved with the
system image.
This avoids storing a large number of types that are usually not needed again.