Don't store source expressions in docstrings
Currently each docstring stores the `Expr` that it documents. With
Documenter.jl I've not found any use for looking at the source
expression since we can get the useful information through reflection
easier than traversing raw expressions.

Instead we store the `Binding` and type signature in place of the
source, which is much more useful when auto-generating docs.

This was undocumented and I've not found anything relying on it within
`METADATA` so should be fine to just remove.

As a bonus removing the source expressions saves 800KB in each of the
`sys.*` files.