Operations between Float16 and Integer now return Float16
Previously we would promote to Float32 leading to subtle type
instabilities when Float16 was used for computations instead of
a pure data-storage format.

By defining `promote_type(Float16, Integer) = Float16` this
incurs an overhead by first converting the `Integer` argument
to `Float16` and then `Float32` for mathematical operations,
that are performed in Float32.