Skip methods(+) doctest
This particular doctest ended up producing a lot of false positives as
it appears the ordering from method can change from commit to commit.
Since this doctest was failing so consistently I set this particular
doctest to be skipped.

Additionally I expanded the sample of + methods shown which was
previously more restricted in an attempt to make this doctest pass more
consistently.