Disable osx and i686-linux travis for now
The queue is getting out of hand. Losing the test coverage on PR's
is unfortunate, but we do have the buildbots that will test things
that get merged to master (though they will do so after the fact,
so this will rely on people keeping a close eye on them to avoid
letting master stay broken for long on those platforms)

Hopefully this will be temporary and we can revert it when things
are quieter.