Move RST 'data' docstrings inline
This moves all `.. data::` docstrings found in the RST docs into
`base/` source files, so all docstrings found in the docs are now
actual docstrings. This will help with the markdown manual conversion.

Minor formatting changes include addition of some fullstops in
several docstrings, and reformatting of the `StackTrace` docstring.