Try ~/.ssh/id_rsa first by default
Only prompt for a different private key. If authentication with the first key fails.
This can be a little annoying if one wants to use a different private key, but the
default one is encrypted, so it queries for a password, without asking for the key first.
However, in that case, one probably wants to use the environment variables or the ssh
agent anyway, so I'm not too concerned.