Bump VERSION to 0.5.0-rc0
See #17418 - this is not entirely ready for release since there are a few
known bugs we need to fix, but it is feature-complete and ready for package
authors to test against. Master is still in bug/doc/test-fix only mode, we
will branch release-0.5 when closer to RC1 which will be more complete.

Note that VERSION will stay at 0.5.0-rc0 until rc1 (or branching, whichever
comes first), builds between rc0 and rc1 will be v"0.5.0-rc0+n".