Fix start_reading(stream, callback)
This fixes two issues with the start_reading that takes a callback,
which currently cannot be used since it throws an error immediately.

The notify_filled change is necessary since the
  search(stream::IOBuffer, delim)
requires the delimiter to be of type UInt8, i.e. specifically
  search(buf::Base.AbstractIOBuffer{Array{UInt8,1}}, delim::UInt8)
  search(buf::Base.AbstractIOBuffer, delim::UInt8)
This error only surfaces when the stream has a readcb attached.

In start_reading it fixes a mismatch between variable names which
makes the call fail with an "UndefVarError: failure_code not defined".