Don't zero `usesshagent`
`usesshagent` is just a flag to determine whether or not to use the agent, so
zeroing it isn't necessary. Even worse, since it only gets a reference to one
of the flag strings (stored in the AST), zeroing them globally changes those
strings, disabling ssh-agent authentication entirely.

Fixes the problem noted in #16041