Refactor libgit2 credential callback
Cleanly separate the SSHCredentials and UserPasswordCredentials to only use each type
when appropriate (the former for SSH keys, the latter for HTTPS for SSH password/keyboard-interactive
authentication). Previously the types would sometimes get confused and SSHCredentials
would be used for userpass authentication, since the field names happen to be a subset
(though they have different meaning, since the `pass` field in SSHCredentials is the
SSH key passphrase while the `pass` in UserPasswordCredentials is the remote user's
password).

Fix a couple issues in the process:
- Allow externally passed in credentials to reach the right place
- Don't skip using the SSH agent if there are two private packages