Only download busybox-w32 when USE_GPL_LIBS is 1
some of the spawn and cmdlineargs tests may fail in a no-GPL
Windows build if executed outside of the build environment
(all of which are GPL themselves...), we could either skip them
or try to get them to work via calling powershell or some other
non-GPL coreutils replacement (toybox? uutils? sbase + now-LPGL cygwin?)