Refactor credentials yet again
Now, we never prompt for credentials unless we consider the credentials invalid (we tried three times and gave up),
(note exceptions are ssh keys which we know have to exist for ssh credentials to work).
Also, if through interactive prompting credentials get changed, they get reset to make sure
that we do the retries. This fixes the original issue of not prompting when a prompt would have
been required as well as being nicer in the face of typos when entering credentials.

Fixes certain clone operations over HTTPS.