Add Gitter to README
The Gitter is a large active community that new users have found lots of help at. I think it would be helpful to have the Gitter right here, because some of the people went to IRC for chat only to find out about the Gitter through the Gitter/IRC bot (and then switched to Gtiter for the code highlighting). I think showing the active community and the help it gives on a README will be important to lower the new-user threshold.