Fix trailing whitespace in code literal
rst apparently requires that the closing double backticks for code literals
appear at the end of a "word", rather than preceded by a space as markdown
allows. This means that ``` ``julia> `` ``` doesn't parse correctly. The
fix is to just remove the trailing space and explicitly mention the need
for a space after the `>`.