Reinstate downloading old versions of winrpm gcc dll's
This mostly reverts #17906 and puts #15521 back in place,
but from a slightly more permanent download location now.

Unfortunately using the latest opensuse copy of the gcc dll's
is causing issues when a cygwin-built Julia tries to load an
opensuse-built libzmq.dll. Using these slightly old gcc 5 versions
seems to fix it. I suspect the difference is which libstdc++ ABI
is being used. GCC 5 has both old and new available, but many distros
had it using the old ABI by default for compatibility. GCC 6 (which
opensuse is using now) is more likely to be using the new ABI by
default.