In cgmemmgr, allow writing permissions on the read page for the debugger
The permissions specified in MapViewOfFile specify the maximum permissions
the file can ever have, which means that even permissions bypass by the
debugger is not allowed to write it. Instead, set the maximum permissions
to RWX and VirtualProtect down to the permissions we need. This is the same
behavior as we use on Mac/Linux.

(cherry picked from commit 4f1ba0cfa517d51652736b47854fedef74430de3)
ref #18083