Prevent repl tests from hanging if homedir() is a symlink
I am testing julia on a machine where my home directory is a
symlink.  Unfortunately, this means the repl tests hang, as they
wait for the homedir to be given at the shell prompt, but instead
the pwd() is printed.  `pwd()` calls `uv_cwd()` which calls
`getcwd()`, which [will always resolve symbolic links](http://stackoverflow.com/a/1545838/1558890),
thus printing something other than `homedir()` on my system.
This changes the repl tests to wait for `realpath(homedir())`
instead of `homedir()`.