Fix relocation verification code
The code needs to use the mangled name in order to find symbols. On OS X,
mangling adds an underscore, while on linux mangling does nothing. Since
this verification code only exists when assertions are turned on, not
many people have run into this (since it's a bug in the verification,
not actually indicative of a problem).

Nevertheless, fixes https://github.com/Keno/Cxx.jl/issues/296.