Fix JIT unwind info registration
For `UNW_INFO_FORMAT_IP_OFFSET`, the `table_len` is the size of the `table_data`
in `unw_word_t` (i.e. `uintptr_t`) for whatever reason.

Fix #19360 (which is only failing on the centos buildbot likely because the
fallback methods somehow do not work on such an old glibc)