Fix cmp for AbstractStrings where indexing is not the same
The original code (from 2010) incorrectly used the index for string a (i) when calling done for string b.
This means that when two different types of abstract strings are compared, if their indexing
is not compatible (for example, where one is indexing by character, and the other by code unit),
then you can get incorrect results, ranging from simply returning that two strings are the equal
when they are not, or getting a BoundsError exception.
This can happen even with common strings with just words with accents.

(cherry picked from commit 4a0798478c6dc9bbe3ba14f996060448efd8bcb1)
ref #19445