Fix completion on custom (non-Dict) Associatives (#19448)
Inference is confused by dict_identifier_key(), which is inherently type
unstable since it calls getfield based on the name of the object
(passed as a string). When completions() is first compiled,
there is only one start() method for Associative objects, so its return type
is assumed for all Associative. This breaks when introducing a custom
Associative like SortedDict, since completions() isn't recompiled because
its input types do not include the dictionary type, and the new start()
methods do not trigger its recompilation (issue #265).

As a temporary solution, move the offending code to a separate function which
takes the dictionary as an argument, so that a specialized version is compiled
and calls the correct methods.
(cherry picked from commit cdabda19a9bf4bac74a6f2d8a0f717d105874354)