empirical performance improvement for type inference
the idea here is that we want to try to avoid accidentally doing a depth-first scan of the AST
but while a breadth-first (dominator) scan is more optimal in terms of avoiding rework,
it is also more expensive to compute

intead, the idea here is that anytime we hit a goto branch,
we instead track whether this is a lower pc waiting for revisit.

by preferring lower pc to larger ones,
we improve the ability for inference to finish completely inferring
all branches of a loop / conditional / try-catch before moving on

this depends on the fact that lowering generally keeps these bloks
fairly well ordered in the AST. excessive use of `@goto` nodes in the user's code
may reduce the effectiveness of this heuristic, but it won't break the algorithm.