Disable /proc/self/mem memmgr on Linux < 4.10
The DirtyCOW exploit mitigation patch in Linux (torvalds/linux@19be0eaff),
introduced a kernel bug, that would cause the kernel to hang if an
application tries to use /proc/<pid>/mem (or ptrace) to bypass page
protections on pages that are backed by transparent huge pages (though only
if the write causes a COW resolution). We make use of this feature in our
cg memory manager on Linux, but since we can't predict whether or not our
mappings will be backed by transparent huge pages, the only safe thing
to do is to fall back to dual maps on kernels that potentially have this
issue. Since the problematic commit is a fix for a high-profile exploit,
it is quite likely that it is included in most stable kernels by now.
I have a patch pending at http://marc.info/?l=linux-mm&m=148359462417378&w=2,
which I expect will fix this in the kernel for 4.10. However, we'll have
to disable the use of /proc/self/mem for all prior kernel versions to
avoid locking up the kernel.