Use compiler intrinsics for byte swapping
There is really no advantage to messing with inline assembly here,
but plenty of disadvantages. The compiler can't optimize this for
whatever domain the data happens to be in (integer vs SSE+ registers),
some of the code violates the C standard, and it generates warnings
on newer gcc platforms (on i686 at least). Just get rid of it, use
the compiler intrinsics, and fall back to generic C instructions for
unknown compilers.