codegen: represent Union types via tagged unions
this covers the code-generator support for representing types on the
stack via a tagged type representation

The function `remark_julia_type` now can transform between any two types.
Instead of asserting when asked to drop type information, it'll instead
cast the object to the requested representation.
It'll make boxes, allocate stack copies, and compute tagged unions as
needed to perform the request.

This means that we can use `remark_julia_type` whenever we want type
information to forcibly change.
For example, storing to a slot requires dropping the intermediate type
information and generating a representation suitable for copying into
the memory location. Loading from a slot has the opposite requirement:
it needs to be able to load the slot via it's declared type,
then possibly to cast the result to add type information that may be
used later (to emit optimizations of other intrinsic function calls).

The primitive routines that need to be able to handle tagged unions are:
- remark-type
- load-local
- store-local
- isa
- is
- emit_typeof
- emit_sizeof
- boxed
- unbox
- specsig (not implemented here yet)

Everything else should be possible to handle in inference by using these
primitives to implement union-splitting.

Additionally, the following will be supported once the type layout changes:
- getfield / setfield

The representation of the tagged-union is as a pair
of < union*, selector > memory addresses.
The selector is fixed as a byte, and will union-tag the first 126 isbits.
It records the one-based depth-first count into the type-union of the
isbits objects inside. An index of zero indicates that the union* is
actually a tagged heap-allocated jl_value_t*,
and needs to treated as normal for a boxed object rather than as a
tagged union.

The `union*` memory region may be allocated at *any* size.
The only constraint is that it is big enough to contain the data
currently specified by `selector`.
It might not be big enough to contain the union of all types that
could be stored there according to the associated Union type field.
Use appropriate care when copying.

this also includes improvements for codegen of isa with mixed types