Fix minor hygiene issue in `@eval` macro (closes #20397)
This is Jameson's suggstion of calling `Core.eval` with
`current_module`; there aren't a lot of situations where this
has an effect on code since all modules by default have an
`eval` function that evaluates code inside of them. However,
you can overwrite this binding with a different value, which
would previously cause `@eval` to fail. The test for this also
serves as a test that `@eval` works in general, with and
without the module argument.