extend mr_empty when the type is supplied
We allow users to call `sum(T::Type, xs)`. Currently, e.g. `sum(Float64, [])` gives an error because the `eltype` of `[]` is `Any` and so inference fails. But the fact that the user supplied a type means that we know the intended `eltype`, so we can use it to set e.g. `sum(Float64, []) = zero(Float64)`.