Use mbedTLS hostname verification for LibGit2 (#20443)
* Use mbedTLS hostname verification for LibGit2

Corrects issues with hostname verification failing when the SSL subjectAltName is not
specified and the CN should be used.

* Add test for mbedTLS verification

* Made mbedTLS verification a seperate patch

* Disable startup-file

* Remove use of which

* Changes from PR review

* Use localhost when hostname isn't set to loopback

* Handle getaddrinfo exception

* Improve warning message

* Add libgit2-mbedtls-hostname.patch dependency

libgit2-mbedtls-hostname patch directly depends on the libgit2-mbedtls
patch. The libgit2-mbedtls-writer-fix patch also modifies the
mbedtls_stream.c file seperate section of the file. To ensure that the
patches are always applied in the same order we'll make the hostname
patch dependent on the writer-fix patch.