hashing of general objects was far too simple (fixes #20744) (#20767)
* hashing of general objects was far too simple (fixes #20744)

* hash_uint(3h - object_id(x)) is a somewhat better hash

If we'd used `3h - object_id(x)` previously, the original issue
would never have come up since `hash(a, hash(b, hash(c, h)))`
wouldn't have produced `3A - 3B + 3C - h` but would instead have given
the value `27h -9A - 3B - C`, which is asymmetrical in `a` and `c`.
Passing the result to `hash_uint` fixes that issue since that function
is non-linear, but defense in depth, right? We may as well do the
same operations in such a way that the passed-through value gets the
most operations done on it.