fix #21118, subtyping issue with diagonal variable with abstract lower bound
In `Tuple{Vararg{T}} where T`, `T` is diagonal. In the case in this issue,
type intersection was giving such a variable an abstract lower bound. That
was wrong, but it also created a type that could not be subtyped properly
since the given bound on the variable contradicted the requirement that it
be concrete. So for now at least, given e.g. `Tuple{T,T} where T>:Real`,
simply drop the diagonal-concrete rule.