Remove manual HTML anchors from README.md.
The `README.md` is currently rendered incorrectly on GitHub since it includes
manual HTML anchors for headlines.  However, since several years, GitHub
automatically includes these anchors.  Considering the anchors in the
README.md have been added some six years ago, the GitHub feature is probably
younger.

This removes all the manual anchors, for which all match the exact headline
pattern anyway, except the following which differ slightly:

- `<a name="Editor-Terminal-Setup"/>`
- `<a name="Platform-Specific-Notes"/>`
- `<a name="Required-Build-Tools-External-Libraries"/>`

If referenced from somewher externally, these would need to be updated.

Links within this README to anchors are also corrected.

The automatically generated anchors on GitHub follow the pattern of replacing
all spaces with `-` and everything being lowercase, e.g. `## This is a Title`
automatically gets the anchor equivalent to `<a name="this-is-a-title"/>`
to be referenced in markdown with `[Link to a title](#this-is-a-title)`.