Re-introduce SSH agent error checking
When re-using credentials like we do with `Pkg.update` we need to allow
`creds.usesshagent == "U"` to allow separate authentications to the same
host to both use the SSH agent. This allows the SSH agent to be used
multiple times when we have multiple private packages.

Re-introducing this code did mean we needed to introduce a way of
disabling the use of the SSH agent for testing purposes.