fix #21397, `Nullable(Tuple)` method error
This was caused by widening a signature in `jl_is_cacheable_sig`
such that we ended up calling a specialization containing an unknown
static parameter value.

This failed when the specialization was created by inference, since
it correctly stored a typevar as the value of the static parameter.
However when the specialization was created by the method caching
logic it accidentally worked, since the static parameter value
was derived before widening.