Make structure of Cmd more obvious in printing
Cmds behave as an array of strings, but for most new users the mental
model they start with is that Cmds are essentially strings. As a result,
people can get easily confused by the interpolation rules around Cmds.
They are quite intuitive when one thinks of them as arrays of strings
(e.g. arrays get expanded, strings get inserted as is), but less so
when thinking of a Cmd as a single string. It seems possible to make
it clearer in printing that there is an underlying structure to the
Cmd by adjusting the way it prints. This one attempt at doing so by
underlining each separate part of the array. Some advantages:
- The underlining makes it clear that there is something going on
  beyong being a regular string
- The underlining gets dropped when copying out from a terminal, so
  there's no trouble with pasting it back into a REPL.
- The underlying structure of clearly visible (each contiguous underline
  is one of the elements of the array of strings).