Make reshape(::Bitarray,...) type stable (#21670)
This works around some of the performance regressions flagged in #20993. The instability itself is not new nor is it the cause of the regression. The trouble is just that there are a few more inlined functions now, each of which needs a dynamic dispatch.

It is worth noting that 0.6 (with this patch) is significantly faster than 0.5 (with this patch). I measure an improvement of ~30%.
(cherry picked from commit 253fa74996c19129dfa326de141e479a9e3150df)