fix #21719, incorrect MethodError
This was caused by incorrectly using a TypeVar as the value of a static
parameter in the case of a compiled specialization whose signature is
not a leaf type. In this case, one slot was specialized on `Function`,
but the slot with the static parameter had a leaf type.

(cherry picked from commit a581292947a758ce7957f3f9caaeebe14c9d284a)
ref #21729