Rename cond(LU) to the helper function _cond1Inf and allow for a norm (#21577)
argument. The old version was broken and inefficient. While it would
be possible to fix just the bug, the performance problem is a consequence
of how condition number estimation works, i.e. a norm estimate is required
and that is not natural to produce from the factorized matrix. Hence,
the change to a helper function.

Fixes #21453