Workaround for FreeBSD linking to outdated system libs
FreeBSD's system libgcc_s declares its GCC version as 4.6, which is too
old to build Julia. Since gfortran doesn't come installed by default,
when it's installed it's done so through a GCC port, which includes its
own libgcc_s. Linking to libgfortran from the port and to the system's
libgcc_s causes versioning issues which wreck the build.

This commit works around this issue by determining the version of GCC
used for gfortran, then linking in the libgcc_s corresponding to that
version everywhere. It's a bit of a kludge, but it gets the job done.
With this change, all libraries which need to link to libgcc_s link to
the proper version.