Clarify precedence of numeric literal coefs over parenthesis
As stated in issue #21798, it was not clear to me that the
precedence of numeric literal coefficients with parenthesized
expressions is also the same as that of unary operators. This
caused confusion with expressions like 6/2(2+1), which results in 1
in julia, but conventionally it should be 9. Hence, the
documentation was slightly modified to make this fact much clear.