Update dates.md
The claim that Julia Base DateTime implements UT (meant is UT1) time would require that the user converts ubiquitous UTC time stamps to UT1 and back, when accuracy requirements are about 1 s or higher. For most users this would be challenging to understand and perform correctly. Moreover, Julia itself gets system time, which is UTC (or UTC plus a timezone offset) with the `now()` function and constructs a DateTime without conversion to UT1, and so does not fulfill the claim. I think, that the claim should be retracted, see also the discussion at https://discourse.julialang.org/t/higher-resolution-datetime-timestamp/3069/35. I have suggested corresponding updates to dates.md