Fix broadcast_indices (#22130)
This fixes a regression introduced in 4f1b479d. broadcast_indices() needs to
be overloaded by packages for custom types, so it cannot be hidden under
_broadcast_indices(). Also, ::Type is incorrect since the method only applies
to scalars.

Make the tests more complex to be closer to actual implementations in packages
so that regressions like this will be noticed in the future.