Drop support for LLVM < 3.9.1 (#22401)
As of the merge of #21888, we no longer support building on LLVM <= 3.9.1.
This is essentially the mechanical cleanup to rip out all code that
we had to support building on older LLVM versions. There may still be
some residual support left in places and of course some things can now
be cleaned up further, but this should get us started.