Add a dead code elimination pass before lowering ptls
Make sure dead ptls references are removed at LLVM IR level and add test for it.
This is mostly for test purpose. LLVM is usually smart enough to remove it later but
it's much harder to test and the present of a dead call in the IR makes many other tests harder.