Deprecate order::Ordering keyword argument to sort* methods
This argument is only used internally and was not supposed to be documented
(which was the case in Julia 0.6). Deprecate it everywhere in the public API,
but keep using it internally for now since the logic is needed to keep
the deprecation working, and for some internal uses like in sortrows() and
sortcols(), which call sortperm() with a Lexicographic ordering.
Remove order for docstrings.

Also deprecate the initialized keyword argument to sort(), which had no effect
and was not documented in Julia 0.6 either.