bin/sha1map.jl: script to map commit (and tag) hashes to tree hashes
This script loads a file in this repo called `bin/sha1map.toml` if it
exists and uses it to map old-style METADATA.jl SHA1 commit hashes
(and the occasional tag hash) to new-style SHA1 tree hashes. If this
file does not exist or doesn't contain a mapping for some uuid/sha1
pair, then it will clone the repo into ~/.julia/upstream (if needed)
and look up the correct mapping from commit/tag to tree hash.