Add convenience function "clean"
Add a convenience function named `clean` that calls `rm` then `add` on its argument in sequence. The justification is that it's a way to "clean" a local package that has been marked as "dirty", and return to the clean master version. This function is based on the recommendation of @greg_plowman in response to a question on [how to revert dirty packages](https://discourse.julialang.org/t/how-to-revert-dirty-packages/2122).
Suggested future improvements: only clean when the package is actually marked as dirty, use `git reset --hard` type functionality, as suggested by @stevengj.