bin/generate: new version data compression format
The compressed format is generically (in TOML):

    ["1.0-2.5.3"]
    # applicable key value pairs
    key1 = "first value"
    key2 = ["some", "array", "values"]

    ["1.2"] # version pattern
    # applicable key value pairs
    key3 = 3.1415

        ["1.2".stuff]
        A = 123
        B = true

This is "decompressed" by accumulating key-value pairs for all
the version range stanzas which a given version number belongs.
If any given key appears in different stanzas with overlapping
version ranges, the file is invalid. This compressed format is
used for all version-specific data in a package registry, e.g.:

    * `dependencies.toml` – mapping dependency names to UUIDs.
    * `compatibility.toml` – mapping dependencies to the sets
      of versions of those dependencies that are compatible.
    * `versions.toml` – mapping version numbers to tree hashes.
    * `targets.toml` – mapping target names to dependencies.

The `versions.toml` file is a degenerate case of the formst in
which each stanza header matches only a single version, and it
can therefore be handled as a simple version-info dictionary,
but it can also be handled using a generic decompressor. There
is something of a bootstrapping issue, however, since one does
need a version or list of versions in order to decompress the
format, so it generally makes sense to parse `versions.toml`
specially and then use the list of version numbers to parse
the other compressed version data files.