Package loading and adding with new config & manifest formats
The `Config.toml` format is largely the same, but now we only have
entries in the `[deps]` stanza for packages that are directly used
by the top-level project. The `Manifest.toml` format now has array
stanzas for each package name, one for each different package by
that name that it includes (often only one). This allows multiple
packages by the same name to be used by a single project. Each of
the packages in `Manifest.toml` has an optional `[deps]` stanza
which, in long form, is the same as that package verion's config
`[deps]` stanza. However, since the most common case is that each
name is unique within a project, the `deps` entry can just map to
an array of dependency names, if the names are all unique within
the `Manifest.toml` file. If _any_ of a stanza's dependencies are
not uniquely named, then it must use the long form, mapping names
to UUIDs in a dictionary-style `[deps]` stanza.