rm operation: remove reverse dependencies and their dependents
The remove operation doesn't actually uninstall any packages, it
works purely on the config and manifest files. The named packages
are the initial "drop set" – any package that depends on these is
also dropped (transitively). This may leave some non-top-level
entries in the manifest file "orphaned" – i.e. no longer present
for any reason, so the operation only keeps packages that are
required by some top-level package in the config file (this is
also computed transitively). This operation only deletes entries
from the config and manifest files – it never changes the version
of any project dependency.