Move LibGit2.prompt to Base.prompt
As part of the supporting EOF update (#23092) we needed to deprecate
the original `LibGit2.prompt` since the return type changed. Since
packages like PkgDev make use of the `prompt` function and the function
itself isn't LibGit2 specific it seemed best to move it into Base. This
allows us to deprecate the original `prompt` without having to change
the function name or modify the parameters.