Add allow_prompt keyword to CredentialPayload
Removes `prompt_if_incorrect` field from `SSHCredentials` and
`UserPasswordCredentials` and replaces it with the `allow_prompt` field
in `CredentialPayload`. Note that allowing prompting is now the default.

Note: Although we can deprecate when people specify a boolean for
the `prompt_if_incorrect` parameter we cannot detect if people were
reliant on prompting being disabled by default.