fixed deprecated syntax; reworded advice
As per @tkelman 's suggestion, I fixed the deprecated syntax in the snippet.  I also reworded the advice at the end of the new text to clarify that parameterization is just one possible way to get performant code, but the key point is that the constructed type should be in the type signature of the arguments of the kernel.