fix possible missing gc-rootedness in REPL test
Previously, we might see the gc finalize the REPL object,
and thus the STDOUT stream reader, too early.
This would leave the STDOUT stream writer with a broken pipe,
resulting in EPIPE errors during the test.

Issue was observed during a Travis CI run for a PR,
and has not been reproduced locally to confirm.

Also took this opportunity to update a few of the constructs
used in the test to more modern Julia APIs.