implement islower, isupper and istitle for String
`islower` and `isupper` were deprecated to a simplistic
behavior which left to be desired; the new functions
check that the predicate is true *only* for letters,
and that there is at least one letter. This is the behavior
of some other languages (e.g. Python) and is inspired by
http://www.unicode.org/L2/L1999/99190.htm.
Also, `islower(lowercase(s))` is now true for many more
strings (false only if s doesn't contain any letter).