Separate out Circle steps and give them names
Without this change, the circle steps have the contents of the commands
smashed together as the step names. This gives each steps a clear name.
To see the commands executed, just expand the step in the web UI and
it's the first thing shown.

The more important change here is that some of the larger steps are
separated out into smaller, more modular steps. This avoids the Circle
web UI being unable to fetch the log, which happens with the enormous
logs generated by running everything in a single step.

[av skip] [bsd skip]