Remove some leagcy gcroot codegen code
The LLVM pass now figures this out automatically, so we no longer
need to put it explicitly into the code. There's more cleanup
to be done here, but I want to take this one step at a time,
since this legacy code may have been hiding issues in the GC
root placement pass.