Make sure we have all required function prototypes before lowering GC intrinsics.
According to the LLVM docs this should be done in doInitialization, but eg.
in the case of stripDeadPrototypes (module pass) -> LowerGCFrame (function pass)
the pass manager initializes all module passes, including the function pass manager
which in turn initializes all function passes, before running any other pass.
This may result in cached prototypes pointing to invalid functions.

Alternatively, we could populate llvm.used with the prototypes and clean it afterwards.