REPL: add an Options type to handle REPL options
It was previously not practical to disable the "align"
feature of backspace: overwriting the '\b' binding with:
`'\b' => (s,o...) -> Base.LineEdit.edit_backspace(s, false)`
worked to a certain extent, but for example it was disabling
the feature that '\b' must get you out of shell/help mode.

A new experimental Options type is introduced here, which is
meant to conveniently allow users to pass options controlling
the REPL behavior. For example:

atreplinit() do repl
    repl.options = Base.REPL.Options(backspace_align = false)
end