Change the behavior of `precompile` depending on current mode
The expected behavior of `precompile` should be doing as much compilation as what would have been
done before actually calling the function.
The previous version of the function does not generate native code so it does not completely
eliminate latency before the function is actually executed for the first time.
Of course, such compilation is also only necessary if it will be used later.
Therefore we can skip some of the work if we know that the result will be discarded.
In summary, with this patch, the behavior of `precompile` will be different depending on which
mode julia is running.

1. During sysimg generation (more specifically, when we are saving llvm IR or object code)
   the behavior will be the same as before since we only need to generate the llvm IR for it
   to be cached in the sysimg.
2. During package precompilation (or otherwise outputting only ji files) do not generate either
   llvm IR or native code since they won't be saved.
3. If we are not saving anything and is therefore in normal mode, generate the native code
   since this is likely what the caller is expecting.